bootp.h: fix size of struct bootp to 300
This was something unset by one of Debian's patches, which I think
was wrong.  The size of this struct should be fixed at 300.  This
struct uses an int32 datatype, which was incorrectly #define'ed to
a long.  On 64-bit systems, this made int32 64-bits, leading to a
packed structure 304 bytes long.  The fix is to int32, not the size
of struct bootp, so revert this change from Debian.